Safari-611.1.10/LayoutTests/fast/forms/image-disconnected-during-parse.html

<!DOCTYPE html>
<form>
    <div id="target">
        <script>
            var d = document.getElementById('target');
            d.parentNode.removeChild(d);
        </script>
        <img name=myimage>
    </div>
</form>
<script src="../../resources/js-test-pre.js"></script>
<script>
description('Image elements should not be associated with forms in different trees');
shouldBeUndefined('document.forms[0].myimage');
</script>
<script src="../../resources/js-test-post.js"></script>
